created a tests folder and moved all test files. Fixed the error in the function dedupe
