+++ Sprint-1/implement/dedupe.js
@@ -1,17 +1,15 @@
-function dedupe() {
-  function dedupe(array) {
-    const result = [];
+function dedupe(array) {
+  const result = [];
 
-    for (let i = 0; i < array.length; i++) {
-      const item = array[i];
-      // Check if item is not already in the result array
-      if (result.indexOf(item) === -1) {
-        result.push(item);
-      }
+  for (let i = 0; i < array.length; i++) {
+    const item = array[i];
+    // Check if item is not already in the result array
+    if (result.indexOf(item) === -1) {
+      result.push(item);
     }
-
-    return result;
   }
+
+  return result;
 }
 
 module.exports = dedupe;
